requirements.txt
appnope==0.1.0
astroid==1.6.5
atomicwrites==1.2.1
attrs==18.2.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
bleach==3.0.2
configparser==3.5.0
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
ipaddress==1.0.22
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.4
more-itertools==4.3.0
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.0
numpy==1.15.2
pandas==0.23.4
pandocfilters==1.4.2
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.5
pluggy==0.8.0
prometheus-client==0.4.2
prompt-toolkit==1.0.15
ptyprocess==0.6.0
py==1.7.0
Pygments==2.2.0
pylint==1.9.3
pytest==3.9.1
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.2
scandir==1.9.0
selenium==3.14.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
wrapt==1.10.11
xvfbwrapper==0.2.9
